+++ backend/attendance/views/instructor_attendance.py
@@ -1,18 +1,15 @@
 #!/usr/bin/env python3
-"""Views for the attendance app - Lecture attendance and instructor fingerprint integration."""
+"""Views for instructor attendance (fingerprint devices and admin dashboard)."""
 
 from rest_framework import generics, status, views
-from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated, IsAdminUser, AllowAny
 from rest_framework.decorators import api_view, permission_classes
 from django.utils import timezone
-from django.db.models import Count, Q
-from django.db import transaction
 from django.shortcuts import get_object_or_404
 from django.core.exceptions import ValidationError
 
-from .models import (
+from attendance.models import (
     InstructorAttendance,
     AttendanceDevice,
     SupervisorSchedule,
@@ -22,15 +19,7 @@ from .models import (
     FingerprintScanLog,
     ScanAction,
 )
-from .models.lecture_attendance import LectureAttendance
-from .permissions import IsAdminOrCourseInstructor
-from .serializers import (
-    # Lecture attendance serializers
-    MarkAttendanceSerializer,
-    LectureAttendanceSerializer,
-    LectureAttendanceDetailSerializer,
-    BulkMarkAttendanceSerializer,
-    # Instructor attendance serializers
+from attendance.serializers import (
     InstructorAttendanceSerializer,
     InstructorAttendanceListSerializer,
     FingerprintCheckInSerializer,
@@ -44,363 +33,6 @@ from .serializers import (
 from users.models import Instructor
 from courses.models import Season
 
-
-# =============================================================================
-# Lecture Attendance Views (for students/children)
-# =============================================================================
-
-class LectureAttendanceView(APIView):
-    """
-    API endpoint to mark attendance for a student or child.
-
-    Only admins and the course instructor can mark attendance.
-
-    POST /api/attendance/lecture/<lecture_id>/mark/
-
-    Request body:
-    {
-        "code": "M64793",
-        "participant_type": "student",
-        "rating": 8,
-        "notes": "Optional notes"
-    }
-    """
-    permission_classes = [IsAdminOrCourseInstructor]
-
-    def post(self, request, lecture_id):
-        """Mark attendance for a student or child using their code."""
-        from courses.models.lecture import Lecture
-
-        # Get the lecture and check permissions
-        try:
-            lecture = Lecture.objects.select_related(
-                'course', 'course__instructor').get(id=lecture_id)
-        except Lecture.DoesNotExist:
-            return Response(
-                {'error': 'Lecture not found.'},
-                status=status.HTTP_404_NOT_FOUND
-            )
-
-        # Check object-level permission (admin or course instructor)
-        if not self.permission_classes[0]().has_object_permission(request, self, lecture):
-            return Response(
-                {'error': 'You do not have permission to mark attendance for this lecture.'},
-                status=status.HTTP_403_FORBIDDEN
-            )
-
-        # Add lecture_id to the request data
-        data = request.data.copy()
-        data['lecture_id'] = lecture_id
-
-        serializer = MarkAttendanceSerializer(
-            data=data,
-            context={'request': request}
-        )
-
-        if not serializer.is_valid():
-            return Response(
-                serializer.errors,
-                status=status.HTTP_400_BAD_REQUEST
-            )
-
-        # Get the validated data and attendance record
-        validated_data = serializer.validated_data
-        attendance = serializer.context['attendance']
-
-        try:
-            with transaction.atomic():
-                # Set rating and notes before marking
-                attendance.rating = validated_data['rating']
-                attendance.notes = validated_data.get('notes', '')
-
-                # Use the model's mark() method to mark as present (attended)
-                attendance.mark(
-                    present=True,
-                    marked_by_user=request.user
-                )
-
-            # Return the updated attendance record
-            response_serializer = LectureAttendanceSerializer(attendance)
-            return Response(
-                {
-                    'message': 'Attendance marked successfully',
-                    'lecture_id': validated_data['lecture_id'],
-                    'attendance': response_serializer.data
-                },
-                status=status.HTTP_200_OK
-            )
-
-        except Exception as e:
-            return Response(
-                {'error': f'Failed to mark attendance: {str(e)}'},
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR
-            )
-
-
-class BulkLectureAttendanceView(APIView):
-    """
-    API endpoint to mark attendance for multiple students/children in bulk.
-
-    Only admins and the course instructor can mark attendance.
-
-    POST /api/attendance/lecture/<lecture_id>/mark-bulk/
-
-    Request body:
-    {
-        "marked_via": "manual",  // or "qr_scan"
-        "attendances": [
-            {
-                "code": "M64793",
-                "participant_type": "student",
-                "rating": 8,
-                "notes": "Good performance",
-                "present": true
-            },
-            {
-                "code": "C12345",
-                "participant_type": "child",
-                "rating": 9,
-                "notes": "Excellent",
-                "present": true
-            }
-        ]
-    }
-
-    Response:
-    {
-        "message": "Bulk attendance marking completed",
-        "lecture_id": 123,
-        "summary": {
-            "total_received": 10,
-            "successful": 8,
-            "failed": 2,
-            "marked_by": "John Doe",
-            "marked_via": "manual",
-            "marked_at": "2026-02-10T10:30:00Z"
-        },
-        "successful_records": [...],
-        "failed_records": [...]
-    }
-    """
-    permission_classes = [IsAdminOrCourseInstructor]
-
-    def post(self, request, lecture_id):
-        """Mark attendance for multiple students/children in bulk."""
-        from courses.models.lecture import Lecture
-
-        # Get the lecture and check permissions
-        try:
-            lecture = Lecture.objects.select_related(
-                'course', 'course__instructor').get(id=lecture_id)
-        except Lecture.DoesNotExist:
-            return Response(
-                {'error': f'Lecture with id {lecture_id} not found.'},
-                status=status.HTTP_404_NOT_FOUND
-            )
-
-        # Check object-level permission (admin or course instructor)
-        if not self.permission_classes[0]().has_object_permission(request, self, lecture):
-            return Response(
-                {'error': 'You do not have permission to mark attendance for this lecture.'},
-                status=status.HTTP_403_FORBIDDEN
-            )
-
-        # Validate the request data
-        serializer = BulkMarkAttendanceSerializer(
-            data=request.data,
-            context={'request': request, 'lecture': lecture}
-        )
-
-        if not serializer.is_valid():
-            return Response(
-                serializer.errors,
-                status=status.HTTP_400_BAD_REQUEST
-            )
-
-        validated_data = serializer.validated_data
-        validated_items = serializer.context.get('validated_items', [])
-        validation_errors = serializer.context.get('validation_errors', [])
-
-        marked_via = validated_data.get('marked_via', 'manual')
-        marked_at = timezone.now()
-
-        # Process all validated attendance records
-        successful_records = []
-        failed_records = []
-
-        try:
-            with transaction.atomic():
-                for item in validated_items:
-                    attendance = item['attendance']
-                    participant = item['participant']
-                    data = item['data']
-
-                    try:
-                        # Set rating and notes
-                        attendance.rating = data['rating']
-                        attendance.notes = data.get('notes', '')
-
-                        # Mark attendance
-                        attendance.mark(
-                            present=data.get('present', True),
-                            marked_by_user=request.user,
-                            marked_via=marked_via
-                        )
-
-                        # Add to successful records
-                        successful_records.append({
-                            'code': data['code'],
-                            'participant_type': data['participant_type'],
-                            'participant_name': (
-                                participant.first_name if hasattr(participant, 'first_name')
-                                else participant.user.get_full_name() if hasattr(participant, 'user')
-                                else 'Unknown'
-                            ),
-                            'rating': data['rating'],
-                            'present': data.get('present', True),
-                            'attendance_id': attendance.id
-                        })
-
-                    except Exception as e:
-                        failed_records.append({
-                            'code': data['code'],
-                            'participant_type': data['participant_type'],
-                            'error': f'Failed to mark attendance: {str(e)}'
-                        })
-
-                # Add validation errors to failed records
-                failed_records.extend(validation_errors)
-
-        except Exception as e:
-            return Response(
-                {'error': f'Transaction failed: {str(e)}'},
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR
-            )
-
-        # Prepare summary
-        total_received = len(validated_data.get('attendances', []))
-        summary = {
-            'total_received': total_received,
-            'successful': len(successful_records),
-            'failed': len(failed_records),
-            'marked_by': request.user.get_full_name() or request.user.username,
-            'marked_via': marked_via,
-            'marked_at': marked_at.isoformat()
-        }
-
-        response_status = status.HTTP_200_OK if len(
-            successful_records) > 0 else status.HTTP_400_BAD_REQUEST
-
-        if len(failed_records) > 0 and len(successful_records) > 0:
-            response_status = status.HTTP_207_MULTI_STATUS
-
-        return Response(
-            {
-                'message': 'Bulk attendance marking completed',
-                'lecture_id': lecture_id,
-                'summary': summary,
-                'successful_records': successful_records,
-                'failed_records': failed_records
-            },
-            status=response_status
-        )
-
-
-class LectureAttendanceDetailView(APIView):
-    """
-    API endpoint to get detailed attendance for a specific lecture.
-
-    Returns all attendance records for the lecture with full participant details
-    including name, image, age, gender, rating, notes, and attendance status.
-
-    Only admins and the course instructor can view this data.
-
-    GET /api/attendance/lecture/<lecture_id>/details/
-
-    Response:
-    {
-        "lecture_id": 123,
-        "lecture_title": "Lecture 1 - Introduction",
-        "course_name": "Quran Memorization",
-        "total_enrolled": 15,
-        "present_count": 12,
-        "absent_count": 3,
-        "attendance_rate": 80.0,
-        "attendances": [
-            {
-                "id": 1,
-                "participant_name": "أحمد",
-                "participant_full_name": "أحمد محمد",
-                "participant_type": "child",
-                "participant_code": "M12345",
-                "participant_image": "https://res.cloudinary.com/.../image.jpg",
-                "participant_age": 12,
-                "participant_gender": "boy",
-                "present": true,
-                "rating": 8,
-                "notes": "ممتاز",
-                "marked_at": "2025-01-15T10:30:00Z"
-            },
-            ...
-        ]
-    }
-    """
-    permission_classes = [IsAdminOrCourseInstructor]
-
-    def get(self, request, lecture_id):
-        """Get detailed attendance records for a lecture."""
-        from courses.models.lecture import Lecture
-
-        # Get the lecture with related data
-        try:
-            lecture = Lecture.objects.select_related(
-                'course', 'course__instructor'
-            ).get(id=lecture_id)
-        except Lecture.DoesNotExist:
-            return Response(
-                {'error': 'Lecture not found.'},
-                status=status.HTTP_404_NOT_FOUND
-            )
-
-        # Check object-level permission (admin or course instructor)
-        if not self.permission_classes[0]().has_object_permission(request, self, lecture):
-            return Response(
-                {'error': 'You do not have permission to view attendance for this lecture.'},
-                status=status.HTTP_403_FORBIDDEN
-            )
-
-        # Get all attendance records for this lecture with related data
-        attendances = LectureAttendance.objects.filter(
-            lecture=lecture
-        ).select_related(
-            'child', 'student', 'student__user', 'marked_by'
-        ).order_by('-present', 'child__first_name', 'student__user__first_name')
-
-        # Calculate statistics
-        total_enrolled = attendances.count()
-        present_count = attendances.filter(present=True).count()
-        absent_count = total_enrolled - present_count
-        attendance_rate = (present_count / total_enrolled *
-                           100) if total_enrolled > 0 else 0
-
-        # Serialize the attendance records
-        serializer = LectureAttendanceDetailSerializer(attendances, many=True)
-
-        return Response({
-            'lecture_id': lecture.id,
-            'lecture_title': lecture.title,
-            'course_name': lecture.course.name if lecture.course else None,
-            'total_enrolled': total_enrolled,
-            'present_count': present_count,
-            'absent_count': absent_count,
-            'attendance_rate': round(attendance_rate, 1),
-            'attendances': serializer.data
-        }, status=status.HTTP_200_OK)
-
-
-# =============================================================================
-# Instructor Attendance Views (for fingerprint devices and admin dashboard)
-# =============================================================================
 
 class DeviceAuthenticationMixin:
     """
